+++ my-app/app/components/employer/index.js
@@ -2,3 +2,4 @@ export * from "./EmployerHeader";
 export * from "./EmployerStats";
 export * from "./JobsTable";
 export * from "./JobRow";
+export * from "./EmployerInsights";